Show pending tx id with spinner
Pending tx is shown after the tx was sent to the network successfully.
Once the corresponding event, confirming the tx, arrives from cnd the
spinner is replaced with the green checkmark.
We only use cnd's confirmation to change the status of the tx at the
moment, confirmation picked up by Ambrosia's wallet are not considered.